refactor: move shouldTriggerOnSuccess
