+++ react/lib/util/api-client.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import _ from 'lodash';
 import config from '../config.json'
-import { getCurrencyTypeFromAddress, isValidCashAddress, isValidXecAddress } from './address';
+import { isValidCashAddress, isValidXecAddress } from './address';
 import {
   Transaction,
   UtxoDetails,
@@ -10,49 +10,6 @@ import {
   Currency,
 } from './types';
 import { isFiat } from './currency';
-import BigNumber from 'bignumber.js';
-import { resolveNumber } from './number';
-
-export const shouldTriggerOnSuccess = (
-  transaction: Transaction,
-  price: number,
-  currency: string,
-  expectedPaymentId?: string,
-  expectedAmount?: BigNumber,
-  expectedOpReturn?: string,
-) => {
-  const {
-    paymentId,
-    rawMessage:rawOpReturn,
-    message,
-    amount, 
-    address } = transaction; 
-  let isAmountValid = true;
-  if(expectedAmount) {
-    const transactionCurrency: Currency = getCurrencyTypeFromAddress(address);
-    if ((transactionCurrency !== currency)) {
-      const value = resolveNumber(amount).times(price)
-      const expectedValue = resolveNumber(amount).times(price)
-      if(value) {
-        isAmountValid = expectedValue.isEqualTo(value);
-      }
-    } else {
-      isAmountValid = expectedAmount.isEqualTo(amount)
-    }
-  } 
-
-  const paymentIdsMatch = expectedPaymentId === paymentId;
-  const isPaymentIdValid = expectedPaymentId ? paymentIdsMatch : true;
-
-  const rawOpReturnIsEmptyOrUndefined = rawOpReturn === '' || rawOpReturn === undefined;
-  const opReturn = rawOpReturnIsEmptyOrUndefined ? message : rawOpReturn
-  const opReturnIsEmptyOrUndefined = opReturn === '' || opReturn === undefined;
-
-  const opReturnsMatch = opReturn === expectedOpReturn;
-  const isOpReturnValid = expectedOpReturn ? opReturnsMatch : opReturnIsEmptyOrUndefined;
-
-  return isAmountValid && isPaymentIdValid && isOpReturnValid;
-};
 
 export const getAddressDetails = async (
   address: string,
